app/src/redisinserter/requirements.txt
redis
google-cloud-storage==1.35.0